+++ R/download.R
@@ -132,7 +132,7 @@ download_file_version <- function(document_uuid,
 #'
 #' @details
 #' This function can be used to read the following data file types:
-#' csv, rds, xlsx.
+#' csv, rds, xlsx, parquet.
 #'
 #' The function works by downloading the file from Objective Connect to a
 #' temporary file and reading it into R. The following functions are used to
@@ -143,6 +143,7 @@ download_file_version <- function(document_uuid,
 #' | csv | \code{readr::read_csv()} |
 #' | rds | \code{readr::read_rds()} |
 #' | xlsx | \code{readxl::read_xlsx()} |
+#' | parquet | \code{nanoparquet::read_parquet()} |
 #'
 #' To check what file type your document is (and thus what function additional
 #' arguments will be passed to), use \code{asset_info()}.